+++ src/main/java/me/daniel/domain/DTO/product/ProductDTO.java
@@ -1,14 +1,17 @@
-package me.daniel.domain.DTO;
+package me.daniel.domain.DTO.product;
 
-public class ProductListDTO {
+public class ProductDTO {
 
     private Integer productNo;
     private String productName;
+    private String productCategory;
     private Integer productPrice;
     private Integer productStock;
+    private String productDetail;
     private String productImage;
+    private Integer productStatus;
 
-    public ProductListDTO() {
+    public ProductDTO() {
 
     }
 
@@ -28,6 +31,14 @@ public class ProductListDTO {
         this.productName = productName;
     }
 
+    public String getProductCategory() {
+        return productCategory;
+    }
+
+    public void setProductCategory(String productCategory) {
+        this.productCategory = productCategory;
+    }
+
     public Integer getProductPrice() {
         return productPrice;
     }
@@ -44,11 +55,27 @@ public class ProductListDTO {
         this.productStock = productStock;
     }
 
+    public String getProductDetail() {
+        return productDetail;
+    }
+
+    public void setProductDetail(String productDetail) {
+        this.productDetail = productDetail;
+    }
+
     public String getProductImage() {
         return productImage;
     }
 
     public void setProductImage(String productImage) {
         this.productImage = productImage;
+    }
+
+    public Integer getProductStatus() {
+        return productStatus;
+    }
+
+    public void setProductStatus(Integer productStatus) {
+        this.productStatus = productStatus;
     }
 }
