+++ src/main/java/me/daniel/domain/DTO/user/UserDTO.java
@@ -1,4 +1,4 @@
-package me.daniel.domain.DTO;
+package me.daniel.domain.DTO.user;
 
 public class UserDTO {
